+++ template_variables/RELEASES.md
@@ -259,3 +259,11 @@ Based on:
 - Speakeasy CLI 1.455.7 (2.480.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.14.0] template_variables
+
+## 2024-12-21 00:28:16
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.459.2 (2.483.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.14.1] template_variables